Fix building on FreeBSD
Closes https://github.com/owncloud/ocis/issues/1217

* Extracted common error methods to `errtypes` package
* Fixed type error on Bavail ( https://golang.org/src/syscall/ztypes_freebsd_arm64.go#L125 )
* Alias ENODATA to ENOATTR

How to test this on Linux:

```
export GOOS=freebsd
make
```

+++ pkg/storage/utils/decomposedfs/decomposedfs_freebsd.go
@@ -16,7 +16,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-// +build !windows,!freebsd
+// +build freebsd
 
 package decomposedfs
 
@@ -28,5 +28,5 @@ func (fs *Decomposedfs) getAvailableSize(path string) (uint64, error) {
 	if err != nil {
 		return 0, err
 	}
-	return stat.Bavail * uint64(stat.Bsize), nil
+	return uint64(stat.Bavail) * uint64(stat.Bsize), nil
 }
